+++ NEWS.md
@@ -4,6 +4,7 @@
 - angle conversion
 - refinement of organism class for modelling
 - organism modeling functions
+- Dolbear's law
 
 # sonicscrewdriver 0.0.5